# CUDA: noncont MMVQ + batched bs1 MUL_MAT_ID