This PR makes the following changes:

* Extend MMVQ to handle noncontiguous and batched inputs. This enables doing the KQ calculation in a single kernel launch when using quantized K and no FlashAttention.
* Add FP32 support to MMV. Basically no change in terms of performance for CUDA. hipBLAS seems to have been very bad for batch size 1 so on my RX 6800 this provides a speedup of almost 5x for FP32 models.
* Extend MMV and MMVQ with support for batch size 1 `MUL_MAT_ID`. This reduces the number of kernel launches needed by a factor equal to the number of used `src0` matrices. It also eliminates the need for a call to `cudaStreamSynchronize` prior to launching the kernels.
* Fix the condition for `MMV_MAX_ROWS`, it was supposed to be `<=` but I accidentally wrote `<`. In my testing this makes a small but measurable difference for Deepseek v2 Lite. Thank you to @jukofyork for pointing this out to me.

<details>
<summary>Performance changes</summary>

| GPU         | Model               | Cache type K | Test   | t/s master | t/s PR | Speedup |
| :---------- | :------------------ | ------------ | :----- | ---------: | -----: |  ------ |
| RX 6800     | deepseek 2 16B Q4_0 | F16          | tg128  |      27.26 |  52.74 |    1.93 |
| RX 6800     | llama 8B Q4_0       | F16          | tg128  |      62.02 |  60.90 |    0.98 |
| RX 6800     | llama 8B Q4_0       | q8_0         | tg4096 |      38.98 |  49.65 |    1.27 |
| RX 6800     | llama 1B all F32    | F16          | tg128  |      16.08 |  76.08 |    4.73 |
| P40         | deepseek 2 16B Q4_0 | F16          | tg128  |      47.78 |  62.87 |    1.32 |
| P40         | llama 8B Q4_0       | F16          | tg128  |      51.02 |  49.91 |    0.98 |
| P40         | llama 8B Q4_0       | q8_0         | tg4096 |      31.96 |  31.71 |    0.99 |
| P40         | llama 1B all F32    | F16          | tg128  |      53.65 |  57.46 |    1.07 |
| 2x P40      | deepseek 2 16B F16  | F16          | tg128  |      34.96 |  37.45 |    1.07 |
| RTX 3090    | deepseek 2 16B Q4_0 | F16          | tg128  |      76.58 | 130.17 |    1.70 |
| RTX 3090    | llama 8B Q4_0       | F16          | tg128  |     140.00 | 138.90 |    0.99 |
| RTX 3090    | llama 8B Q4_0       | q8_0         | tg4096 |      86.82 |  93.21 |    1.07 |
| RTX 3090    | llama 1B all F32    | F16          | tg128  |     151.58 | 152.11 |    1.00 |
| RTX 4090    | deepseek 2 16B Q4_0 | F16          | tg128  |      55.84 | 123.55 |    2.21 |
| RTX 4090    | llama 8B Q4_0       | F16          | tg128  |     158.67 | 157.21 |    0.99 |
| RTX 4090    | llama 8B Q4_0       | q8_0         | tg4096 |     104.08 | 123.94 |    1.19 |
| RTX 4090    | llama 1B all F32    | F16          | tg128  |     167.65 | 168.72 |    1.01 |
| 2x RTX 4090 | deepseek 2 16B F16  | F16          | tg128  |      69.80 |  87.62 |    1.26 |

</details>

Notes:

* The performance of Deepseek v2 Lite is better on my RTX 3090 than my RTX 4090. I think there is still significant kernel launch overhead so the difference comes from my 3090 being paired with a faster CPU.
* There is a small performance regression for non-MoE models; I think even disregarding MoE this is a worthwhile tradeoff since user code now has more flexibility regarding tensor shapes for the CUDA backend.
* The approach I used for this PR only works for batch size 1 where the arithmetic intensity is terrible anyways because you can use each loaded weight only once. I have an idea that I think will work well for MMQ but would make it necessary to compile twice as many template specializations. For floating-point data I have so far not been able to think of a good solution that doesn't involve writing my own GEMM kernel (which would be a lot of work).