@slaren 2025-04-18
I am testing this, but with deepseek2 the ROPE is being run on the CPU, which destroys performance. If I change the `supports_op` to force it to run, the performance is much better. Also, the exception that disables CUDA graphs with MUL_MAT_ID could be removed now, which would improve performance further. With both of these fixes the performance that I get on my 3090 Ti is similar to what you report here, otherwise it is much lower.