@JohannesGaessler 2025-04-18
I converted my gguf file from safetensors after the recent deepseek changes, @jukofyork does this explain the discrepancy?