@slaren 2025-04-18
I tried with a freshly converted model from https://huggingface.co/deepseek-ai/DeepSeek-Coder-V2-Lite-Instruct, but the ROPE is still being run on the CPU, so that doesn't seem to be the problem.