@JohannesGaessler 2025-04-19
Oops, I was only looking at the runtimes of CUDA kernels and forgot to check whether some ggml ops still run on the CPU 😅. Non-contiguous inputs for RoPE are already implemented, but the check in `ggml_backend_cuda_device_supports_op` was wrong. CUDA graphs can be enabled for `MUL_MAT_ID` if the batch size is 1 because for other batch sizes a workaround is still needed (but I think for batch sizes > 1 CUDA graphs are currently disabled anyways). With those two fixes the performance on my systems now is:

| GPU         | Model               | Test   | t/s master | t/s PR old | t/s PR new | Speedup vs. master | Speedup vs. PR old |
| :---------- | :------------------ | :----- | ---------: | ---------: | ---------: | ------------------ | ------------------ |
| RX 6800     | deepseek 2 16B Q4_0 | pp512  |     565.29 | -          |     637.31 |               1.13 | -                  |
| RX 6800     | deepseek 2 16B Q4_0 | tg128  |      27.26 | 52.74      |      71.95 |               2.64 | 1.36               |
| P40         | deepseek 2 16B Q4_0 | pp512  |     727.03 | -          |     841.71 |               1.16 | -                  |
| P40         | deepseek 2 16B Q4_0 | tg128  |      47.78 | 62.87      |      71.13 |               1.49 | 1.13               |
| RTX 3090    | deepseek 2 16B Q4_0 | pp512  |    1885.98 | -          |    2344.70 |               1.24 | -                  |
| RTX 3090    | deepseek 2 16B Q4_0 | tg128  |      76.58 | 130.17     |     184.17 |               2.40 | 1.41               |
| RTX 4090    | deepseek 2 16B Q4_0 | pp512  |    2882.90 | -          |    3570.49 |               1.24 | -                  |
| RTX 4090    | deepseek 2 16B Q4_0 | tg128  |      55.84 | 123.55     |     223.27 |               4.00 | 1.81               |