@slaren 2025-04-19
Comparison with master on Windows (WSL, RTX 3090 Ti):

| Model              | Test   |   t/s master |   t/s cuda-moe-mmv-2 |   Speedup |
|:-------------------|:-------|-------------:|---------------------:|----------:|
| deepseek2 16B Q4_0 | tg128  |        29.13 |               190.94 |      6.56 |
| llama 8x7B Q3_K_S  | tg128  |        45.88 |                70.85 |      1.54 |